add utils.py in anatomist_tools
